Modify all files
Modified README to add pseudo code to handle restarts.
Modified index.html slightly to find the restart button.
Modified style.css just to hide restart button from the start.
Modified script.js to add functionality to the restart button. Users can now restart the game.

This is the extent of the changes for now. I may come back and make the game look nice or fix script.js because it is probably not great.