Modify index.html and script.js
Added some more UI to the game in index.html so that users can see the result of the round (win, loss, or tie) and the result of the game (win or loss).
Added functionality to script.js to display the result of each round and to the game and then display those on the page.